AI_MOCK_INTERVIEW/requirements.txt
streamlit
openai
elevenlabs
PyPDF2
PymuPDF
python-docx
pytesseract
Pillow
python-dotenv
sounddevice
soundfile
googletrans==4.0.0-rc1
SpeechRecognition
numpy
pandas
langchain>=0.1.0
langchain_openai>=0.1.0
tiktoken
tqdm
python-dotenv
requests
elevenlabs
python-docx
